Split core into two parts
